Revert "Topic/mnaveau/isolate yaml parser"
