--- agimus_controller/agimus_controller/ocp/ocp_force_feedback.yaml
@@ -1,92 +1,0 @@
----
-running_model:
-  class: IAMSoftContactAugmented
-  differential:
-    class: DAMSoftContactAugmentedFwdDynamics
-    Kp: [6000, 6000, 6000]
-    Kv: [100, 100, 100]
-    frame_id: panda_hand_tcp
-    oPc: [0.0, 0.0, -0.073915]
-    with_gravity_torque_reg: true
-    enabled_directions: [true, true, true]
-    ref: LOCAL_WORLD_ALIGNED
-    cost_ref: LOCAL_WORLD_ALIGNED
-    costs:
-    - name: state_reg
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelState
-    - name: frame_rot_tracking_tool
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelFrameRotationStatic
-          frame_id: panda_hand_tcp
-    - name: frame_pose_tracking_tool
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelFrameTranslationStatic
-          frame_id: panda_hand_tcp
-
-terminal_model:
-  class: IAMSoftContactAugmented
-  differential:
-    class: DAMSoftContactAugmentedFwdDynamics
-    Kp: [6000, 6000, 6000]
-    Kv: [100, 100, 100]
-    frame_id: panda_hand_tcp
-    oPc: [0.0, 0.0, -0.073915]
-    with_gravity_torque_reg: false
-    enabled_directions: [true, true, true]
-    ref: LOCAL_WORLD_ALIGNED
-    cost_ref: LOCAL_WORLD_ALIGNED
-    costs:
-    - name: state_reg
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelState
-    - name: frame_rot_tracking_tool
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelFrameRotationStatic
-          frame_id: panda_hand_tcp
-    - name: frame_pose_tracking_tool
-      update: true
-      weight: 1.0
-      cost:
-        class: CostModelResidual
-        activation:
-          class: ActivationModelWeightedQuad
-          weights: 1.0
-        residual:
-          class: ResidualModelFrameTranslationStatic
-          frame_id: panda_hand_tcp
