--- agimus_controller/tests/test_ocp_force_feedback.py
@@ -1,140 +1,0 @@
-import numpy as np
-from pathlib import Path
-import example_robot_data
-import unittest
-import force_feedback_mpc  # noqa: F401
-import pinocchio
-
-from agimus_controller.trajectory import (
-    TrajectoryPoint,
-    TrajectoryPointWeights,
-    WeightedTrajectoryPoint,
-)
-
-from agimus_controller.ocp.ocp_croco_generic_force_feedback import (
-    OCPCrocoForceFeedbackGeneric,
-    get_globals,
-)
-from agimus_controller.factory import ocp_yaml_parser
-from agimus_controller.factory.robot_model import RobotModels, RobotModelParameters
-from agimus_controller.ocp_param_base import DTFactorsNSeq
-from agimus_controller.ocp_param_base import OCPParamsBaseCroco
-
-
-class OCPCrocoGenericTest(unittest.TestCase):
-    def setUp(self):
-        ### Set force_feedback_mpc to globals
-        ocp_yaml_parser.add_modules(get_globals())
-
-        ### LOAD ROBOT
-        robot = example_robot_data.load("panda")
-        urdf_path = Path(robot.urdf)
-        srdf_path = Path(robot.urdf.replace("urdf", "srdf"))
-        urdf_meshes_dir = urdf_path.parent.parent.parent.parent.parent
-        free_flyer = False
-        locked_joint_names = ["panda_finger_joint1", "panda_finger_joint2"]
-        reduced_nq = robot.model.nq - len(locked_joint_names)
-        moving_joint_names = set(robot.model.names) - set(
-            locked_joint_names + ["universe"]
-        )
-        q0 = np.zeros(robot.model.nq)
-        armature = np.full(reduced_nq, 0.1)
-
-        # Store shared initial parameters
-        self.params = RobotModelParameters(
-            q0=q0,
-            free_flyer=free_flyer,
-            moving_joint_names=moving_joint_names,
-            robot_urdf=urdf_path,
-            srdf=srdf_path,
-            urdf_meshes_dir=urdf_meshes_dir,
-            collision_as_capsule=True,
-            self_collision=False,
-            armature=armature,
-        )
-
-        self.robot_models = RobotModels(self.params)
-
-        # OCP parameters
-        dt = 0.05
-        horizon_size = 200
-        solver_iters = 100
-        callbacks = False
-
-        self._ocp_params = OCPParamsBaseCroco(
-            dt=dt,
-            horizon_size=horizon_size,
-            dt_factor_n_seq=DTFactorsNSeq(factors=[1], n_steps=[horizon_size]),
-            solver_iters=solver_iters,
-            callbacks=callbacks,
-        )
-
-    def test_ocp_solution(self):
-        # Set initial state
-        q0 = pinocchio.neutral(self.robot_models.robot_model)
-        ee_pose = pinocchio.SE3(np.eye(3), np.array([0.5, 0.2, 0.5]))
-
-        n_states = self._ocp_params.n_controls + 1
-
-        state_warmstart = [
-            np.concatenate(
-                (q0, np.zeros(self.robot_models.robot_model.nv), np.zeros(3))
-            )
-        ] * n_states
-        control_warmstart = [
-            np.zeros(self.robot_models.robot_model.nv)
-        ] * self._ocp_params.n_controls
-        trajectory_points = [
-            WeightedTrajectoryPoint(
-                TrajectoryPoint(
-                    robot_configuration=q0,
-                    robot_velocity=np.zeros(self.robot_models.robot_model.nv),
-                    robot_effort=np.zeros(self.robot_models.robot_model.nv),
-                    forces={"panda_hand_tcp": pinocchio.Force.Zero()},
-                    end_effector_poses={"panda_hand_tcp": ee_pose},
-                ),
-                TrajectoryPointWeights(
-                    w_robot_configuration=0.01
-                    * np.ones(self.robot_models.robot_model.nq),
-                    w_robot_velocity=0.01 * np.ones(self.robot_models.robot_model.nv),
-                    w_robot_effort=0.0001 * np.ones(self.robot_models.robot_model.nv),
-                    w_forces={"panda_hand_tcp": np.zeros(6)},
-                    w_end_effector_poses={"panda_hand_tcp": (1e3 * np.ones(6))},
-                ),
-            )
-        ] * n_states
-
-        # Solve OCP
-        self._state_reg = np.concatenate(
-            (q0, np.zeros(self.robot_models.robot_model.nv), np.zeros(3))
-        )
-        ocp_definition_file = ocp_yaml_parser.get_default_yaml_file(
-            "ocp_force_feedback.yaml"
-        )
-        self._ocp = OCPCrocoForceFeedbackGeneric(
-            self.robot_models, self._ocp_params, yaml_file=ocp_definition_file
-        )
-        self._ocp.set_reference_weighted_trajectory(trajectory_points)
-        self._ocp.solve(self._state_reg, state_warmstart, control_warmstart)
-
-        data = self.robot_models.robot_model.createData()
-        pinocchio.framesForwardKinematics(
-            self.robot_models.robot_model,
-            data,
-            self._ocp.ocp_results.states[-1][: self.robot_models.robot_model.nq],
-        )
-        # Test that the last position of the end-effector is close to the target
-        self.assertAlmostEqual(
-            np.linalg.norm(
-                data.oMf[
-                    self.robot_models.robot_model.getFrameId("panda_hand_tcp")
-                ].translation
-                - ee_pose.translation
-            ),
-            0.0,
-            places=1,
-        )
-
-
-if __name__ == "__main__":
-    unittest.main()
